Android Icon, Fragments in the Drawer Menu, Recycler View
Our drawer menu is now all functional with the implementation of Fragments, their design and their UX.
Updated My Resorts, which needs more work to be done but now has a Recycler View, which will show the users Favorite Ski Resorts on a simple click.
Designed an Application Icon using Adobe XD, to match our theme, and our means of the application.